Simplify Unix implementation for System.Drawing.Printing types.
* Previously, PrintingServices had an implementation which was hidden
  behind a layer of indirection, one for Windows and one for Unix.
  Realistically, the types involved with this were purely static, and the
  indirection was never useful at runtime. By that, I mean that the
  implementation was known at build time -- it is either Windows or Unix.
* I've removed the layer of indirection and deleted all of mono's
  Win32-specific code from the System.Drawing.Printing namespace. Win32
  builds should just use our code from the .NET Framework.
